@@ -5,10 +5,10 @@ author: SlickTech <dev@slicktech.org>
 homepage: https://github.com/SlickTech/svgpath
 
 environment:
-  sdk: '>=1.24.3 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
-dependencies:
-  dart2_constant: ^1.0.0
+# dependencies:
+  # dart2_constant: ^1.0.0
 
 dev_dependencies:
   path: ^1.0.0
