@@ -2,8 +2,9 @@
 // is governed by a MIT license that can be found in the LICENSE file.
 
 import 'dart:math';
+import 'dart:math' as math;
 
-import 'package:dart2_constant/math.dart' as math;
+// import 'package:m/math.dart' as math;
 
 // Convert an arc to a sequence of cubic bézier curves
 
@@ -32,7 +33,7 @@ num vectorAngle(num ux, num uy, num vx, num vy) {
 // see http://www.w3.org/TR/SVG11/implnote.html#ArcImplementationNotes
 //
 // Return [cx, cy, theta1, delta_theta]
-List<num> getArcCenter(num x1, num y1, num x2, num y2, num fa, num fs, num rx,
+List<num> getArcCenter(num x1, num y1, num x2, num y2, num? fa, num? fs, num rx,
     num ry, num sin_phi, num cos_phi) {
   // Step 1.
   //
@@ -40,11 +41,11 @@ List<num> getArcCenter(num x1, num y1, num x2, num y2, num fa, num fs, num rx,
   // points. After that, rotate it to line up ellipse axes with coordinate
   // axes.
   //
-  var x1p =  cos_phi*(x1-x2)/2 + sin_phi*(y1-y2)/2;
-  var y1p = -sin_phi*(x1-x2)/2 + cos_phi*(y1-y2)/2;
+  var x1p = cos_phi * (x1 - x2) / 2 + sin_phi * (y1 - y2) / 2;
+  var y1p = -sin_phi * (x1 - x2) / 2 + cos_phi * (y1 - y2) / 2;
 
-  var rx_sq  =  rx * rx;
-  var ry_sq  =  ry * ry;
+  var rx_sq = rx * rx;
+  var ry_sq = ry * ry;
   var x1p_sq = x1p * x1p;
   var y1p_sq = y1p * y1p;
 
@@ -53,7 +54,7 @@ List<num> getArcCenter(num x1, num y1, num x2, num y2, num fa, num fs, num rx,
   // Compute coordinates of the centre of this ellipse (cx', cy')
   // in the new coordinate system.
   //
-  var radicant = (rx_sq * ry_sq) - (rx_sq * y1p_sq) - (ry_sq * x1p_sq);
+  num radicant = (rx_sq * ry_sq) - (rx_sq * y1p_sq) - (ry_sq * x1p_sq);
 
   if (radicant < 0) {
     // due to rounding errors it might be e.g. -1.3877787807814457e-17
@@ -63,23 +64,23 @@ List<num> getArcCenter(num x1, num y1, num x2, num y2, num fa, num fs, num rx,
   radicant /= (rx_sq * y1p_sq) + (ry_sq * x1p_sq);
   radicant = sqrt(radicant) * (fa == fs ? -1 : 1);
 
-  var cxp = radicant *  rx/ry * y1p;
-  var cyp = radicant * -ry/rx * x1p;
+  var cxp = radicant * rx / ry * y1p;
+  var cyp = radicant * -ry / rx * x1p;
 
   // Step 3.
   //
   // Transform back to get centre coordinates (cx, cy) in the original
   // coordinate system.
   //
-  var cx = cos_phi*cxp - sin_phi*cyp + (x1+x2)/2;
-  var cy = sin_phi*cxp + cos_phi*cyp + (y1+y2)/2;
+  num cx = cos_phi * cxp - sin_phi * cyp + (x1 + x2) / 2;
+  num cy = sin_phi * cxp + cos_phi * cyp + (y1 + y2) / 2;
 
   // Step 4.
   //
   // Compute angles (theta1, delta_theta).
   //
-  var v1x =  (x1p - cxp) / rx;
-  var v1y =  (y1p - cyp) / ry;
+  var v1x = (x1p - cxp) / rx;
+  var v1y = (y1p - cyp) / ry;
   var v2x = (-x1p - cxp) / rx;
   var v2y = (-y1p - cyp) / ry;
 
@@ -93,7 +94,7 @@ List<num> getArcCenter(num x1, num y1, num x2, num y2, num fa, num fs, num rx,
     delta_theta += tau;
   }
 
-  return [ cx, cy, theta1, delta_theta ];
+  return [cx, cy, theta1, delta_theta];
 }
 
 //
@@ -101,24 +102,34 @@ List<num> getArcCenter(num x1, num y1, num x2, num y2, num fa, num fs, num rx,
 // see http://math.stackexchange.com/questions/873224
 //
 List<num> approximateUnitArc(theta1, delta_theta) {
-  var alpha = 4/3 * tan(delta_theta/4);
+  var alpha = 4 / 3 * tan(delta_theta / 4);
 
   var x1 = cos(theta1);
   var y1 = sin(theta1);
   var x2 = cos(theta1 + delta_theta);
   var y2 = sin(theta1 + delta_theta);
 
-  return [ x1, y1, x1 - y1*alpha, y1 + x1*alpha, x2 + y2*alpha, y2 - x2*alpha, x2, y2 ];
+  return [
+    x1,
+    y1,
+    x1 - y1 * alpha,
+    y1 + x1 * alpha,
+    x2 + y2 * alpha,
+    y2 - x2 * alpha,
+    x2,
+    y2
+  ];
 }
 
-List<List<num>> a2c(num x1, num y1, num x2, num y2, num fa, num fs, num rx, num ry, phi) {
+List<List<num>> a2c(
+    num x1, num y1, num x2, num y2, num? fa, num? fs, num? rx, num? ry, phi) {
   var sin_phi = sin(phi * tau / 360);
   var cos_phi = cos(phi * tau / 360);
 
   // Make sure radii are valid
   //
-  var x1p =  cos_phi*(x1-x2)/2 + sin_phi*(y1-y2)/2;
-  var y1p = -sin_phi*(x1-x2)/2 + cos_phi*(y1-y2)/2;
+  var x1p = cos_phi * (x1 - x2) / 2 + sin_phi * (y1 - y2) / 2;
+  var y1p = -sin_phi * (x1 - x2) / 2 + cos_phi * (y1 - y2) / 2;
 
   if (x1p == 0 && y1p == 0) {
     // we're asked to draw line to itself
@@ -130,18 +141,16 @@ List<List<num>> a2c(num x1, num y1, num x2, num y2, num fa, num fs, num rx, num 
     return [];
   }
 
-
   // Compensate out-of-range radii
   //
-  rx = rx.abs();
-  ry = ry.abs();
+  rx = rx!.abs();
+  ry = ry!.abs();
 
   var lambda = (x1p * x1p) / (rx * rx) + (y1p * y1p) / (ry * ry);
   if (lambda > 1) {
     rx *= sqrt(lambda);
     ry *= sqrt(lambda);
   }
-
 
   // Get center parameters (cx, cy, theta1, delta_theta)
   //
@@ -171,12 +180,12 @@ List<List<num>> a2c(num x1, num y1, num x2, num y2, num fa, num fs, num rx, num 
       var y = curve[i + 1];
 
       // scale
-      x *= rx;
-      y *= ry;
+      x *= rx!;
+      y *= ry!;
 
       // rotate
-      var xp = cos_phi*x - sin_phi*y;
-      var yp = sin_phi*x + cos_phi*y;
+      var xp = cos_phi * x - sin_phi * y;
+      var yp = sin_phi * x + cos_phi * y;
 
       // translate
       curve[i] = xp + cc[0];
@@ -186,4 +195,3 @@ List<List<num>> a2c(num x1, num y1, num x2, num y2, num fa, num fs, num rx, num 
     return curve;
   }).toList();
 }
-
